refactored to remove useless extra file
